balanceservice/requirements.txt
Flask
Flask-SQLAlchemy
Flask-JWT-Extended
pyodbc
python-dotenv
Flask-Bcrypt